[FE] feat: 회원가입 text filed 컴포넌트 분리
